Removed extraneous functions from featherstone.c
