+++ src/helpers.c
@@ -332,3 +332,28 @@ void set_texture(char *name, unsigned int tex, unsigned int shader, int unit) {
   glUniform1i(glGetUniformLocation(shader, name), unit);
 }
 
+/*
+  The following functions are helpers to concisely print out the vector and
+  matrix types provided by cglm
+*/
+void print_vec3(vec3 v) {
+  printf("%f %f %f\n", v[0], v[1], v[2]);
+}
+
+void print_vec4(vec4 v) {
+  printf("%f %f %f %f\n", v[0], v[1], v[2], v[3]);
+}
+
+void print_mat3(mat3 m) {
+  for (int i = 0; i < 3; i++) {
+    printf("|%f %f %f|\n", m[0][i], m[1][i], m[2][i]);
+  }
+}
+
+void print_mat4(mat4 m) {
+  for (int i = 0; i < 4; i++) {
+    printf("|%f %f %f %f|\n", m[0][i], m[1][i], m[2][i], m[3][i]);
+  }
+}
+
+
