+++ src/math/spatial_algebra.c
@@ -1,3 +1,4 @@
+#include <stdio.h>
 #include <math/spatial_algebra.h>
 
 void mat6_zero(mat6 m) {
@@ -481,3 +482,20 @@ void vec3_singular_cross(vec3 a, mat3 dest) {
   dest[2][1] = -a[0];
   dest[2][2] = 0.0;
 }
+
+void print_vec6(vec6 v) {
+  printf("%f %f %f %f %f %f\n", v[0], v[1], v[2], v[3], v[4], v[5]);
+}
+
+void print_mat6(mat6 m) {
+  for (int i = 0; i < 6; i++) {
+    printf("|%f %f %f %f %f %f|\n",
+           m[0][i],
+           m[1][i],
+           m[2][i],
+           m[3][i],
+           m[4][i],
+           m[5][i]);
+  }
+}
+
